+++ feincms3_language_sites/middleware.py
@@ -5,7 +5,7 @@ from django.utils import translation
 
 
 def site_middleware(get_response):
-    from .models import site_for_host
+    from feincms3_language_sites.models import site_for_host
 
     def middleware(request):
         request.site = site_for_host(request.get_host())