+++ tests/testapp/urls.py
@@ -3,6 +3,7 @@ from django.contrib import admin
 from django.http import HttpResponse, HttpResponseRedirect
 from django.shortcuts import render
 from django.urls import include, path, re_path
+
 from testapp import views